update: files restructured and jwt change to cookie
