+++ controllers/auth/googleCallback.js
@@ -1,0 +1,6 @@
+const { jwtSignToken } = require("@/helpers/jwt");
+
+exports.googleCallback = (req, res, next) => {
+  const jwtToken = jwtSignToken({ email: req.user.email });
+  return res.json({ userData: req.user.toClient(), jwtToken });
+};
