+++ controllers/user/index.js
@@ -1,0 +1,9 @@
+const { updateProfile } = require("./updateProfile");
+const { profilePicture } = require("./profilePicture");
+const { searchUser } = require("./searchUser");
+
+module.exports = {
+  updateProfile,
+  profilePicture,
+  searchUser,
+};
