+++ helpers/redis.js
@@ -1,4 +1,4 @@
-const redisClient = require("../configs/redis.js");
+const redisClient = require("@/configs/redis.js");
 
 const checkIsUserOnline = async (userId) => {
   const user = await redisClient.get(`chat-cast:userId:${userId}`);
