+++ package.json
@@ -20,6 +20,7 @@
     "body-parser": "^1.20.2",
     "bullmq": "^5.7.8",
     "connect-redis": "^7.1.1",
+    "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "dayjs": "^1.11.10",
     "dotenv": "^16.4.5",
@@ -30,6 +31,7 @@
     "helmet": "^7.1.0",
     "ioredis": "^5.4.1",
     "jsonwebtoken": "^9.0.2",
+    "module-alias": "^2.2.3",
     "mongoose": "^8.3.2",
     "nodemailer": "^6.9.13",
     "passport": "^0.7.0",
@@ -44,5 +46,8 @@
   "optionalDependencies": {
     "bufferutil": "^4.0.8",
     "utf-8-validate": "^6.0.4"
+  },
+  "_moduleAliases": {
+    "@": "./"
   }
 }
